requirements.txt
django
djangorestframework
django-filters
markdown